BigInt conversion in DB

+++ db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import MetaData, String, Integer, Column, Text, DateTime, Boolean, ForeignKey
+from sqlalchemy import MetaData, String, Integer, Column, BigInteger, ForeignKey
 from sqlalchemy.orm import Session
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import relationship, declarative_base
@@ -12,32 +12,32 @@ DataBase = declarative_base()
 class Robot(AuthBase):
     __tablename__ = "robots"
     
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     telegram_ids = relationship("TelegramID", backref="robot")
-    serial = Column(Integer, unique=True)
+    serial = Column(BigInteger, unique=True)
     location = Column(String, nullable=True, default=None)
 
 
 class TelegramID(AuthBase):
     __tablename__ = 'telegram_ids'
 
-    id = Column(Integer, primary_key=True)
-    group_id = Column(Integer)
-    robot_id = Column(Integer, ForeignKey('robots.id'))
+    id = Column(BigInteger, primary_key=True)
+    group_id = Column(BigInteger)
+    robot_id = Column(BigInteger, ForeignKey('robots.id'))
 
 
 class RobotDataDB(DataBase):
     __tablename__ = "robot_data"
-    serial = Column(Integer, primary_key=True, nullable=False)
+    serial = Column(BigInteger, primary_key=True, nullable=False)
 
-    total_bags = Column(Integer, nullable=True)
-    shift_bags = Column(Integer, nullable=True)
+    total_bags = Column(BigInteger, nullable=True)
+    shift_bags = Column(BigInteger, nullable=True)
 
-    left_status = Column(Integer, nullable=True)
-    left_pallet_status = Column(Integer, nullable=True)
+    left_status = Column(BigInteger, nullable=True)
+    left_pallet_status = Column(BigInteger, nullable=True)
 
-    right_status = Column(Integer, nullable=True)
-    right_pallet_status = Column(Integer, nullable=True)
+    right_status = Column(BigInteger, nullable=True)
+    right_pallet_status = Column(BigInteger, nullable=True)
 
     connected = Column(Integer, nullable=True)
